+++ app/src/main/java/nextstep/signup/ui/signup/PasswordField.kt
@@ -1,0 +1,43 @@
+package nextstep.signup.ui.signup
+
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextField
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.text.input.PasswordVisualTransformation
+import androidx.compose.ui.tooling.preview.Preview
+
+
+@Composable
+fun PasswordField(
+    password: String,
+    onPasswordChange: (String) -> Unit
+) {
+    TextField(
+        value = password,
+        onValueChange = { onPasswordChange(it) },
+        modifier = Modifier.fillMaxWidth(),
+        label = { Text(text = "Password") },
+        placeholder = { Text(text = "Password") },
+        visualTransformation = PasswordVisualTransformation(),
+        keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Password)
+    )
+}
+
+
+@Preview(showBackground = true)
+@Composable
+private fun PasswordFieldPreview() {
+    var password by remember { mutableStateOf("") }
+    PasswordField(
+        password = password,
+        onPasswordChange = { password = it }
+    )
+}
